Implemented Buy Transactions.

+++ ggc-core/src/ggc/BreakdownTransaction.java
@@ -5,7 +5,7 @@ public class BreakdownTransaction extends Transaction{
     private int _dateDeadline;
     private String _recipe;
 
-    public BreakdownTransaction(int id, int paymentDate, String productId, String partnerId, int quantity, int price, String recipe){       
+    public BreakdownTransaction(int id, int paymentDate, String productId, String partnerId, int quantity, float price, String recipe){       
         super(id, paymentDate, productId, partnerId, quantity, price);
         _recipe = recipe; 
     }
